Drop mixin
