+++ moriarty/matrix/operator_/operator_.py
@@ -18,7 +18,6 @@ from moriarty.matrix.operator_.autoscaler import (
     get_autoscaler_manager,
 )
 from moriarty.matrix.operator_.dbutils import get_db_session
-from moriarty.matrix.operator_.mixin import EndpointMixin
 from moriarty.matrix.operator_.orm import AutoscalerORM, EndpointORM, InferenceLogORM
 from moriarty.matrix.operator_.params import (
     ContainerScope,
@@ -34,6 +33,31 @@ from moriarty.matrix.operator_.spawner import plugin
 from moriarty.matrix.operator_.spawner.manager import get_spawner
 from moriarty.sidecar.params import MatrixCallback
 from moriarty.sidecar.producer import JobProducer
+
+
+class EndpointMixin:
+    session: AsyncSession
+
+    async def get_endpoint_orm(self, endpoint_name: str) -> EndpointORM | None:
+        return (
+            await self.session.execute(
+                select(EndpointORM).where(EndpointORM.endpoint_name == endpoint_name)
+            )
+        ).scalar_one_or_none()
+
+    async def get_avaliable_endpoints(self) -> list[str]:
+        endpoint_names = (
+            (
+                await self.session.execute(
+                    select(EndpointORM.endpoint_name).where(
+                        EndpointORM.replicas > 0,
+                    )
+                )
+            )
+            .scalars()
+            .all()
+        )
+        return endpoint_names
 
 
 def get_bridger(
